+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Menus/VRCQuestToolsMenus.cs
@@ -59,7 +59,8 @@ namespace KRT.VRCQuestTools.Menus
         internal static class GameObjectMenuPaths
         {
             private const string MenuPrefix = "GameObject/VRCQuestTools/";
-            internal const string ConvertAvatarForQuest = MenuPrefix + "Convert Avatar For Android";
+            internal const string ConvertAvatarForQuest = MenuPrefix + "Convert Avatar for Android";
+            internal const string NdmfManualBakeWithAndroidSettings = MenuPrefix + "[NDMF] Manual Bake with Android Settings";
             internal const string RemovePhysBones = MenuPrefix + "Remove PhysBones";
             internal const string RemoveMissingComponents = MenuPrefix + "Remove Missing Components";
             internal const string RemoveUnsupportedComponents = MenuPrefix + "Remove Unsupported Components";
@@ -69,7 +70,8 @@ namespace KRT.VRCQuestTools.Menus
         internal enum GameObjectMenuPriorities : int
         {
             GameObjectConvertAvatarForQuest = 30,
-            GameObjectRemovePhysBones,
+            GameObjectNdmfManualBakeWithAndroidSettings,
+            GameObjectRemovePhysBones = 130,
             GameObjectRemoveMissingComponents,
             GameObjectRemoveUnsupportedComponents,
             GameObjectRemoveAllVertexColors,